Improve health check and debugability
We suspect that our setup is leaking connections and that is
forcing us to restart haproxy regularly, but because the container
usually has no network it makes it super hard to debug in prod.

This PR add the most basic tools we would need to debug issues.

It also forces a restart when the safe docker is not able to forward
a _ping request.